requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
python-binance
joblib